# backend/swagger.json
{
  "swagger": "2.0",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "info": {
    "version": "1.0.1",
    "title": "Tic-tac-toe"
  },
  "schemes": [
    "http"
  ],
  "definitions": {
    "game": {
      "type": "object",
      "description": "A game object",
      "required": [
        "board"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid",
          "description": "The game's UUID, read-only, generated by the server. The client can not POST or PUT this.",
          "readOnly": true
        },
        "board": {
          "type": "string",
          "description": "The board state",
          "example": "XO--X--OX"
        },
        "status": {
          "type": "string",
          "readOnly": true,
          "description": "The game status, read-only, the client can not POST or PUT this",
          "enum": [
            "RUNNING",
            "X_WON",
            "O_WON",
            "DRAW"
          ]
        }
      }
    }
  },
  "paths": {
    "/api/v1/games": {
      "get": {
        "description": "Get all games.",
        "responses": {
          "200": {
            "description": "Successful response, returns an array of games, returns an empty array if no users found",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/game"
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "Start a new game.",
        "parameters": [
          {
            "name": "game",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/game"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Game successfully started",
            "headers": {
              "Location": {
                "type": "string",
                "description": "URL of the started game"
              }
            },
            "schema": {
              "type": "object",
              "properties": {
                "location": {
                  "type": "string",
                  "description": "URL of the started game"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "reason": {
                  "type": "string",
                  "description": "Why the game failed to start"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/games/{game_id}": {
      "get": {
        "description": "Get a game.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "description": "Game id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response, returns the game",
            "schema": {
              "$ref": "#/definitions/game"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "description": "Post a new move to a game.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "description": "Game id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "game",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/game"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Move successfully registered, also provide backend's response move in response",
            "schema": {
              "$ref": "#/definitions/game"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "reason": {
                  "type": "string",
                  "description": "Why the game failed to update"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "description": "Delete a game.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "description": "Game id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Game successfully deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}